modified logic for delete and added missing relationships

+++ src/main/java/utcn/stackoverflow/stackoverflow/service/UserService.java
@@ -4,9 +4,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import utcn.stackoverflow.stackoverflow.dto.UserDTO;
+import utcn.stackoverflow.stackoverflow.entity.Content;
 import utcn.stackoverflow.stackoverflow.entity.User;
 import utcn.stackoverflow.stackoverflow.repository.UserRepository;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,14 +25,16 @@ public class UserService {
             return null;
         }
 
-        return users.stream().map(user -> new UserDTO(user.getFirstName(), user.getLastName())).toList();
+        return users.stream().map(user -> new UserDTO(user.getUserId(), user.getFirstName(), user.getLastName())).toList();
     }
 
     public UserDTO getUserById(Long cnp) {
         Optional<User> user = userRepository.findByUserId(cnp);
 
         if (user.isPresent()) {
-            return new UserDTO(user.get().getFirstName(), user.get().getLastName());
+            User userPresent = user.get();
+
+            return new UserDTO(userPresent.getUserId(), userPresent.getFirstName(), userPresent.getLastName());
         } else {
             return null;
         }
@@ -40,14 +44,27 @@ public class UserService {
         Optional<User> user = userRepository.findByEmail(email);
 
         if (user.isPresent()) {
-            return new UserDTO(user.get().getFirstName(), user.get().getLastName());
+            User userPresent = user.get();
+
+            return new UserDTO(userPresent.getUserId(), userPresent.getFirstName(), userPresent.getLastName());
         } else {
             return null;
         }
     }
 
     public long deleteUserById(Long id) {
-        return userRepository.deleteByUserId(id);
+        Optional<User> user = userRepository.findByUserId(id);
+
+        if (user.isPresent()) {
+            User userFound = user.get();
+            List<Content> contentList = new ArrayList<>(userFound.getContentList());
+
+            for (Content content : contentList) {
+                userFound.removeContent(content);
+            }
+            return userRepository.deleteByUserId(id);
+        } else
+            return -1;
     }
 
     public User saveUser(User user) {
